[ADD] 강의아이템 컴포넌트 분리
